+++ R/windRose.R
@@ -693,8 +693,16 @@ windRose <- function(mydata, ws = "ws", wd = "wd", ws2 = NA, wd2 = NA,
     if (length(id > 0)) tmp[id] <- 360
     tmp <- table(tmp) ## number of sectors spanned
     vars <- grep("Interval[1-9]", names(results)) ## the frequencies, without any calms
+    
+    # number of rows in data where wd != -999
+    n_data <- nrow(filter(results, wd != -999))
+    
+    if (n_data > 0) {
+      
     results[results[["wd"]] != -999, vars] <-
       results[results[["wd"]] != -999, vars] * mean(tmp) / tmp
+    }
+    
     return(results)
   }