+++ R/checkPrep.R
@@ -138,8 +138,7 @@ checkPrep <- function(mydata, Names, type, remove.calm = TRUE, remove.neg = TRUE
     ## daylight saving time can cause terrible problems - best avoided!!
 
     if (any(dst(mydata$date))) {
-      warning("Detected data with Daylight Saving Time, converting to UTC/GMT")
-      mydata$date <- lubridate::force_tz(mydata$date, tzone = "GMT")
+      message("Detected data with Daylight Saving Time.")
     }
   }