+++ NEWS.md
@@ -1,6 +1,7 @@
 # openair 2.8-99
 
 - fix bug in `windRose` where whole period is calm
+- add optimisation to `polarCluster` to speed up clustering through option `pamonce = 3`. This should not appreciably affect results.
 
 ## openair 2.8-1