+++ NEWS.md
@@ -3,6 +3,7 @@
 - fix bug in `windRose` where whole period is calm
 - add optimisation to `polarCluster` to speed up clustering through option `pamonce = 3`. This should not appreciably affect results.
 - fix strange bug in `aqStats` due to `lubridate` time zone issue.
+- fix bug in `TaylorDiagram` when group was present.
 
 ## openair 2.8-1